Fix PATH overwrite issue
Setting an environment variable with magic_shell has the disadvantage
that variables with the same name get overwritten, as the same file
in /etc/profile.d is created. Using a ruby_block has also issues, it
fails if the file to change is not already present, due to the two
phases of the Chef run, see http://tinyurl.com/qylfxjj.
A pending pull request of magic_shell addresses this issue:
https://github.com/customink-webops/magic_shell/pull/18